imx-ml-demos/imx-face-recognition-demo/requirements.txt
opencv-python
pillow
face_recognition
